fix: file field inheritance
